requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
Django==4.2.25
djangorestframework==3.15.2
idna==3.11
psycopg2-binary==2.9.10
python-dotenv==1.0.1
requests==2.32.4
sqlparse==0.5.3
typing_extensions==4.13.2
urllib3==2.2.3
